Task ip

+++ resume/tasks.py
@@ -40,16 +40,17 @@ def create_visit_task():
     db2 = redis.Redis(host='localhost', port=6379, db=2, decode_responses=True)
     all_keys = db2.keys()
     create_list = []
-    for key in all_keys:
-        val = json.loads(db2.get(key))
-        create_list.append(
-            UniqueIP(
-                ip_address=key,
-                path_client=val['path_client'],
-                info_client=val['info_client'],
+    if all_keys:
+        for key in all_keys:
+            val = json.loads(db2.get(key))
+            create_list.append(
+                UniqueIP(
+                    ip_address=key,
+                    path_client=val['path_client'],
+                    info_client=val['info_client'],
+                )
             )
+        UniqueIP.objects.bulk_create(
+            create_list
         )
-    UniqueIP.objects.bulk_create(
-        create_list
-    )
-    db2.flushdb()
+        db2.flushdb()
